requirements.txt
TexSoup
GitPython
PyMuPDF
